requirements.txt
anyio==4.0.0
certifi==2025.1.31
Flask==2.2.5
httpx==0.25.2
idna==3.7
sniffio==1.3.1
python-dotenv==1.0.1
requests==2.28.2
werkzeug==2.2.3
python-telegram-bot[webhooks]==20.7
